fix: apply sol style

+++ contracts/yearnv1/contracts/IVault.sol
@@ -5,8 +5,12 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 
 interface IVault is IERC20 {
     function token() external view returns (address);
+
     function claimInsurance() external; // NOTE: Only yDelegatedVault implements this
-    function getPricePerFullShare() external view returns (uint);
-    function deposit(uint) external;
-    function withdraw(uint) external;
+
+    function getPricePerFullShare() external view returns (uint256);
+
+    function deposit(uint256) external;
+
+    function withdraw(uint256) external;
 }
