+++ contracts/yearnv1/contracts/vaults/yDelegatedVault.sol
@@ -21,41 +21,46 @@ contract yDelegatedVault is ERC20 {
 
     address public governance;
     address public controller;
-    uint public insurance;
-    uint public healthFactor = 4;
+    uint256 public insurance;
+    uint256 public healthFactor = 4;
 
-    uint public ltv = 65;
-    uint public max = 100;
+    uint256 public ltv = 65;
+    uint256 public max = 100;
 
-    address public constant aave = address(0x24a42fD28C976A61Df5D00D0599C34c4f90748c8);
+    address public constant aave =
+        address(0x24a42fD28C976A61Df5D00D0599C34c4f90748c8);
 
-    constructor (address _token, address _controller) public ERC20(
-        string(abi.encodePacked("yearn ", ERC20(_token).name())),
-        string(abi.encodePacked("y", ERC20(_token).symbol()))
-    ) {
+    constructor(address _token, address _controller)
+        public
+        ERC20(
+            string(abi.encodePacked("yearn ", ERC20(_token).name())),
+            string(abi.encodePacked("y", ERC20(_token).symbol()))
+        )
+    {
         _setupDecimals(ERC20(_token).decimals());
         token = IERC20(_token);
         governance = msg.sender;
         controller = _controller;
     }
 
-    function debt() public view returns (uint) {
+    function debt() public view returns (uint256) {
         address _reserve = IController(controller).want(address(this));
-        (,uint currentBorrowBalance,,,,,,,,) = Aave(getAave()).getUserReserveData(_reserve, address(this));
+        (, uint256 currentBorrowBalance, , , , , , , , ) =
+            Aave(getAave()).getUserReserveData(_reserve, address(this));
         return currentBorrowBalance;
     }
 
-    function credit() public view returns (uint) {
+    function credit() public view returns (uint256) {
         return IController(controller).balanceOf(address(this));
     }
 
     // % of tokens locked and cannot be withdrawn per user
     // this is impermanent locked, unless the debt out accrues the strategy
-    function locked() public view returns (uint) {
+    function locked() public view returns (uint256) {
         return credit().mul(1e18).div(debt());
     }
 
-    function debtShare(address _lp) public view returns (uint) {
+    function debtShare(address _lp) public view returns (uint256) {
         return debt().mul(balanceOf(_lp)).mul(totalSupply());
     }
 
@@ -67,7 +72,7 @@ contract yDelegatedVault is ERC20 {
         return LendingPoolAddressesProvider(aave).getLendingPoolCore();
     }
 
-    function setHealthFactor(uint _hf) external {
+    function setHealthFactor(uint256 _hf) external {
         require(msg.sender == governance, "!governance");
         healthFactor = _hf;
     }
@@ -76,7 +81,7 @@ contract yDelegatedVault is ERC20 {
         Aave(getAave()).setUserUseReserveAsCollateral(underlying(), true);
     }
 
-    function repay(address reserve, uint amount) public  {
+    function repay(address reserve, uint256 amount) public {
         // Required for certain stable coins (USDT for example)
         IERC20(reserve).approve(address(getAaveCore()), 0);
         IERC20(reserve).approve(address(getAaveCore()), amount);
@@ -85,25 +90,25 @@ contract yDelegatedVault is ERC20 {
 
     function repayAll() public {
         address _reserve = reserve();
-        uint _amount = IERC20(_reserve).balanceOf(address(this));
+        uint256 _amount = IERC20(_reserve).balanceOf(address(this));
         repay(_reserve, _amount);
     }
 
     // Used to swap any borrowed reserve over the debt limit to liquidate to 'token'
-    function harvest(address reserve, uint amount) external {
+    function harvest(address reserve, uint256 amount) external {
         require(msg.sender == controller, "!controller");
         require(reserve != address(token), "token");
         IERC20(reserve).safeTransfer(controller, amount);
     }
 
     // Ignore insurance fund for balance calculations
-    function balance() public view returns (uint) {
+    function balance() public view returns (uint256) {
         return token.balanceOf(address(this)).sub(insurance);
     }
 
     function setGovernance(address _governance) external {
-      require(msg.sender == governance, "!governance");
-      governance = _governance;
+        require(msg.sender == governance, "!governance");
+        governance = _governance;
     }
 
     function setController(address _controller) external {
@@ -115,7 +120,7 @@ contract yDelegatedVault is ERC20 {
         return LendingPoolAddressesProvider(aave).getPriceOracle();
     }
 
-    function getReservePriceETH(address reserve) public view returns (uint) {
+    function getReservePriceETH(address reserve) public view returns (uint256) {
         return Oracle(getAaveOracle()).getAssetPrice(reserve);
     }
 
@@ -123,19 +128,25 @@ contract yDelegatedVault is ERC20 {
         return (over() > 0);
     }
 
-    function over() public view returns (uint) {
+    function over() public view returns (uint256) {
         over(0);
     }
 
-    function getUnderlyingPriceETH(uint _amount) public view returns (uint) {
-        _amount = _amount.mul(getUnderlyingPrice()).div(uint(10)**ERC20(address(token)).decimals()); // Calculate the amount we are withdrawing in ETH
+    function getUnderlyingPriceETH(uint256 _amount)
+        public
+        view
+        returns (uint256)
+    {
+        _amount = _amount.mul(getUnderlyingPrice()).div(
+            uint256(10)**ERC20(address(token)).decimals()
+        ); // Calculate the amount we are withdrawing in ETH
         return _amount.mul(ltv).div(max).div(healthFactor);
     }
 
-    function over(uint _amount) public view returns (uint) {
+    function over(uint256 _amount) public view returns (uint256) {
         address _reserve = reserve();
-        uint _eth = getUnderlyingPriceETH(_amount);
-        (uint _maxSafeETH,uint _totalBorrowsETH,) = maxSafeETH();
+        uint256 _eth = getUnderlyingPriceETH(_amount);
+        (uint256 _maxSafeETH, uint256 _totalBorrowsETH, ) = maxSafeETH();
         _maxSafeETH = _maxSafeETH.mul(105).div(100); // 5% buffer so we don't go into a earn/rebalance loop
         if (_eth > _maxSafeETH) {
             _maxSafeETH = 0;
@@ -143,16 +154,21 @@ contract yDelegatedVault is ERC20 {
             _maxSafeETH = _maxSafeETH.sub(_eth); // Add the ETH we are withdrawing
         }
         if (_maxSafeETH < _totalBorrowsETH) {
-            uint _over = _totalBorrowsETH.mul(_totalBorrowsETH.sub(_maxSafeETH)).div(_totalBorrowsETH);
-            _over = _over.mul(uint(10)**ERC20(_reserve).decimals()).div(getReservePrice());
+            uint256 _over =
+                _totalBorrowsETH.mul(_totalBorrowsETH.sub(_maxSafeETH)).div(
+                    _totalBorrowsETH
+                );
+            _over = _over.mul(uint256(10)**ERC20(_reserve).decimals()).div(
+                getReservePrice()
+            );
             return _over;
         } else {
             return 0;
         }
     }
 
-    function _rebalance(uint _amount) internal {
-        uint _over = over(_amount);
+    function _rebalance(uint256 _amount) internal {
+        uint256 _over = over(_amount);
         if (_over > 0) {
             if (_over > credit()) {
                 _over = credit();
@@ -174,52 +190,75 @@ contract yDelegatedVault is ERC20 {
         insurance = 0;
     }
 
-    function maxSafeETH() public view returns (uint maxBorrowsETH, uint totalBorrowsETH, uint availableBorrowsETH) {
-         (,,uint _totalBorrowsETH,,uint _availableBorrowsETH,,,) = Aave(getAave()).getUserAccountData(address(this));
-        uint _maxBorrowETH = (_totalBorrowsETH.add(_availableBorrowsETH));
-        return (_maxBorrowETH.div(healthFactor), _totalBorrowsETH, _availableBorrowsETH);
+    function maxSafeETH()
+        public
+        view
+        returns (
+            uint256 maxBorrowsETH,
+            uint256 totalBorrowsETH,
+            uint256 availableBorrowsETH
+        )
+    {
+        (, , uint256 _totalBorrowsETH, , uint256 _availableBorrowsETH, , , ) =
+            Aave(getAave()).getUserAccountData(address(this));
+        uint256 _maxBorrowETH = (_totalBorrowsETH.add(_availableBorrowsETH));
+        return (
+            _maxBorrowETH.div(healthFactor),
+            _totalBorrowsETH,
+            _availableBorrowsETH
+        );
     }
 
     function shouldBorrow() external view returns (bool) {
         return (availableToBorrowReserve() > 0);
     }
 
-    function availableToBorrowETH() public view returns (uint) {
-        (uint _maxSafeETH,uint _totalBorrowsETH, uint _availableBorrowsETH) = maxSafeETH();
+    function availableToBorrowETH() public view returns (uint256) {
+        (
+            uint256 _maxSafeETH,
+            uint256 _totalBorrowsETH,
+            uint256 _availableBorrowsETH
+        ) = maxSafeETH();
         _maxSafeETH = _maxSafeETH.mul(95).div(100); // 5% buffer so we don't go into a earn/rebalance loop
         if (_maxSafeETH > _totalBorrowsETH) {
-            return _availableBorrowsETH.mul(_maxSafeETH.sub(_totalBorrowsETH)).div(_availableBorrowsETH);
+            return
+                _availableBorrowsETH.mul(_maxSafeETH.sub(_totalBorrowsETH)).div(
+                    _availableBorrowsETH
+                );
         } else {
             return 0;
         }
     }
 
-    function availableToBorrowReserve() public view returns (uint) {
+    function availableToBorrowReserve() public view returns (uint256) {
         address _reserve = reserve();
-        uint _available = availableToBorrowETH();
+        uint256 _available = availableToBorrowETH();
         if (_available > 0) {
-            return _available.mul(uint(10)**ERC20(_reserve).decimals()).div(getReservePrice());
+            return
+                _available.mul(uint256(10)**ERC20(_reserve).decimals()).div(
+                    getReservePrice()
+                );
         } else {
             return 0;
         }
     }
 
-    function getReservePrice() public view returns (uint) {
+    function getReservePrice() public view returns (uint256) {
         return getReservePriceETH(reserve());
     }
 
-    function getUnderlyingPrice() public view returns (uint) {
+    function getUnderlyingPrice() public view returns (uint256) {
         return getReservePriceETH(underlying());
     }
 
     function earn() external {
         address _reserve = reserve();
-        uint _borrow = availableToBorrowReserve();
+        uint256 _borrow = availableToBorrowReserve();
         if (_borrow > 0) {
             Aave(getAave()).borrow(_reserve, _borrow, 2, 7);
         }
         //rebalance here
-        uint _balance = IERC20(_reserve).balanceOf(address(this));
+        uint256 _balance = IERC20(_reserve).balanceOf(address(this));
         if (_balance > 0) {
             IERC20(_reserve).safeTransfer(controller, _balance);
             IController(controller).earn(address(this), _balance);
@@ -230,20 +269,19 @@ contract yDelegatedVault is ERC20 {
         deposit(token.balanceOf(msg.sender));
     }
 
-    function deposit(uint _amount) public {
-        uint _pool = balance();
+    function deposit(uint256 _amount) public {
+        uint256 _pool = balance();
         token.safeTransferFrom(msg.sender, address(this), _amount);
 
         // 0.5% of deposits go into an insurance fund incase of negative profits to protect withdrawals
         // At a 4 health factor, this is a -2% position
-        uint _insurance = _amount.mul(50).div(10000);
+        uint256 _insurance = _amount.mul(50).div(10000);
         _amount = _amount.sub(_insurance);
         insurance = insurance.add(_insurance);
 
-
         //Controller can claim insurance to liquidate to cover interest
 
-        uint shares = 0;
+        uint256 shares = 0;
         if (totalSupply() == 0) {
             shares = _amount;
         } else {
@@ -265,14 +303,14 @@ contract yDelegatedVault is ERC20 {
     }
 
     // No rebalance implementation for lower fees and faster swaps
-    function withdraw(uint _shares) public {
-        uint r = (balance().mul(_shares)).div(totalSupply());
+    function withdraw(uint256 _shares) public {
+        uint256 r = (balance().mul(_shares)).div(totalSupply());
         _burn(msg.sender, _shares);
         _rebalance(r);
         token.safeTransfer(msg.sender, r);
     }
 
-    function getPricePerFullShare() external view returns (uint) {
+    function getPricePerFullShare() external view returns (uint256) {
         return balance().mul(1e18).div(totalSupply());
     }
 }
