+++ contracts/yearnv1/interfaces/Aave.sol
@@ -2,42 +2,60 @@
 pragma solidity ^0.6.2;
 
 interface Aave {
-    function borrow(address _reserve, uint _amount, uint _interestRateModel, uint16 _referralCode) external;
-    function setUserUseReserveAsCollateral(address _reserve, bool _useAsCollateral) external;
-    function repay(address _reserve, uint _amount, address payable _onBehalfOf) external payable;
+    function borrow(
+        address _reserve,
+        uint256 _amount,
+        uint256 _interestRateModel,
+        uint16 _referralCode
+    ) external;
+
+    function setUserUseReserveAsCollateral(
+        address _reserve,
+        bool _useAsCollateral
+    ) external;
+
+    function repay(
+        address _reserve,
+        uint256 _amount,
+        address payable _onBehalfOf
+    ) external payable;
+
     function getUserAccountData(address _user)
         external
         view
         returns (
-            uint totalLiquidityETH,
-            uint totalCollateralETH,
-            uint totalBorrowsETH,
-            uint totalFeesETH,
-            uint availableBorrowsETH,
-            uint currentLiquidationThreshold,
-            uint ltv,
-            uint healthFactor
+            uint256 totalLiquidityETH,
+            uint256 totalCollateralETH,
+            uint256 totalBorrowsETH,
+            uint256 totalFeesETH,
+            uint256 availableBorrowsETH,
+            uint256 currentLiquidationThreshold,
+            uint256 ltv,
+            uint256 healthFactor
         );
+
     function getUserReserveData(address _reserve, address _user)
         external
         view
         returns (
-            uint currentATokenBalance,
-            uint currentBorrowBalance,
-            uint principalBorrowBalance,
-            uint borrowRateMode,
-            uint borrowRate,
-            uint liquidityRate,
-            uint originationFee,
-            uint variableBorrowIndex,
-            uint lastUpdateTimestamp,
+            uint256 currentATokenBalance,
+            uint256 currentBorrowBalance,
+            uint256 principalBorrowBalance,
+            uint256 borrowRateMode,
+            uint256 borrowRate,
+            uint256 liquidityRate,
+            uint256 originationFee,
+            uint256 variableBorrowIndex,
+            uint256 lastUpdateTimestamp,
             bool usageAsCollateralEnabled
         );
 }
 
 interface LendingPoolAddressesProvider {
     function getLendingPool() external view returns (address);
+
     function getLendingPoolCore() external view returns (address);
+
     function getPriceOracle() external view returns (address);
 }
 
@@ -46,6 +64,7 @@ interface AaveToken {
 }
 
 interface Oracle {
-    function getAssetPrice(address reserve) external view returns (uint);
-    function latestAnswer() external view returns (uint);
+    function getAssetPrice(address reserve) external view returns (uint256);
+
+    function latestAnswer() external view returns (uint256);
 }
