+++ contracts/yearnv1/interfaces/Converter.sol
@@ -2,5 +2,5 @@
 pragma solidity ^0.6.2;
 
 interface Converter {
-    function convert(address) external returns (uint);
+    function convert(address) external returns (uint256);
 }
