+++ contracts/yearnv1/interfaces/DForce.sol
@@ -2,16 +2,23 @@
 pragma solidity ^0.6.2;
 
 interface dRewards {
-    function withdraw(uint) external;
+    function withdraw(uint256) external;
+
     function getReward() external;
-    function stake(uint) external;
-    function balanceOf(address) external view returns (uint);
+
+    function stake(uint256) external;
+
+    function balanceOf(address) external view returns (uint256);
+
     function exit() external;
 }
 
 interface dERC20 {
-  function mint(address, uint256) external;
-  function redeem(address, uint) external;
-  function getTokenBalance(address) external view returns (uint);
-  function getExchangeRate() external view returns (uint);
+    function mint(address, uint256) external;
+
+    function redeem(address, uint256) external;
+
+    function getTokenBalance(address) external view returns (uint256);
+
+    function getExchangeRate() external view returns (uint256);
 }
