+++ contracts/yearnv1/interfaces/Gauge.sol
@@ -2,9 +2,11 @@
 pragma solidity ^0.6.2;
 
 interface Gauge {
-    function deposit(uint) external;
-    function balanceOf(address) external view returns (uint);
-    function withdraw(uint) external;
+    function deposit(uint256) external;
+
+    function balanceOf(address) external view returns (uint256);
+
+    function withdraw(uint256) external;
 }
 
 interface Mintr {
